+++ VS/CHANGELOG.md
@@ -1,3 +1,7 @@
+**(Next Version) Changes:**
+
+* Added Effective Health
+
 **1.1.2 Changes:**
 
 * Fixed holdout zone time remaining not working for clients in multiplayer